Commit of first (broken) render of a sphere

+++ src/render/simple_sampler.cpp
@@ -7,7 +7,7 @@ using rgb_t = rgb_t<std::uint8_t>;
 
 void simple_sampler::render(const scene::scene& s, const film& f, std::ostream& os)
 {
-    os << "P3\n" << f.horizontal_pixels << " " << f.vertical_pixels << '\n' << std::numeric_limits<rgb_t::value_type>::max() << '\n';
+    os << "P3\n" << f.horizontal_pixels << " " << f.vertical_pixels << '\n' << static_cast<int>(std::numeric_limits<rgb_t::value_type>::max()) << '\n';
 
     for (std::size_t i = 0; i < f.horizontal_pixels; i++)
         for (std::size_t j = 0; j < f.vertical_pixels; j++)
